Optimize Shiki language loading
After deploying the project to Vercel, we found that it took long time
to open the view page, causing timeout error. After some experiments, it
is the first time loading of the Shiki highlighter that caused the
issue.

To solve the issue, we choose to only load the necessary languages
instead of the all.

Next, we may use ISR to compile all code view pages on compile time and
only load languages which are being using, or, we can load and save the
highlighter on the compile time?